add items to rooms
