+++ util/room_generator.py
@@ -2,6 +2,7 @@
 from adventure.models import Player, Room, Creature
 from adventure.room_templates import room_arrays_dict, MAP_WIDTH, MAP_HEIGHT, BLOCKED_CHARS, EMPTY_CHARS, DOOR_CHARS
 from util.blocks import place_block, place_door
+from util.add_items import add_creatures
 import json
 import random
 import os
@@ -9,27 +10,7 @@ import os
 print(BLOCKED_CHARS)
 # ƺ
 # generate creatures in rooms
-def add_creatures(target_room):
-    import json
-    import random
-    from adventure.models import Player, Room, Creature
-    from adventure.room_templates import room_arrays_dict, MAP_WIDTH, MAP_HEIGHT, BLOCKED_CHARS, EMPTY_CHARS, DOOR_CHARS
-    MIN_CREATURES = 1
-    MAX_CREATURES = 3
-    for i in range(3):
-        # get room and find a legal starting point
-        room_array = json.loads(target_room.room_array)
-        target_char = '█' # dummy
-        while target_char in (BLOCKED_CHARS or DOOR_CHARS):
-            x = random.randrange(1, len(room_array[0]))
-            y = random.randrange(1, len(room_array))
-            target_char = room_array[y][x]
-        creature = Creature(name='Gorgon',
-                            x = x,
-                            y = y,
-                            currentRoom = target_room.id
-                            )
-        creature.save()
+
 
 wd = os.getcwd()
 Room.objects.all().delete()
@@ -100,17 +81,6 @@ while len(rooms)> 4:
         print(select_room.title, "<>", new_room.title)
         rooms.remove(new_room)
 
-
-        # new_room_array = json.loads(new_room.room_array)
-        # select_room.room_array = json.dumps(place_door(select_room_array, select_direction))
-        # new_room.room_array = json.dumps(place_door(new_room_array, opposite_direction))
-        # select_room.connectRooms(new_room,select_direction)
-        # new_room.connectRooms(select_room,opposite_direction)
-        # # rooms.remove(select_room)
-        # # save
-        # select_room.save()
-        # new_room.save()
-
     # chain to the next room in While loop
     select_room = new_room
 
@@ -127,6 +97,9 @@ for i in range(remaining):
     new_room = rooms[i]
 
 place_door(new_room, new_direction, exit_room)
+
+
+
 
 # add creatures to all rooms
 for target_room in Room.objects.all():
